requirements.txt
beautifulsoup4==4.9.3
selenium==3.141.0
python-dotenv==0.17.1
